Definition of the class hierarchy that models the tasks executable by the bots.
The classes in this hierarchy model the tasks that a bot executes to carry out the tasks of the protocol assigned to the bot. These tasks can be simulated through a timeout (doing nothing), or they can be run through a sequence of programmed instructions for the bot, or they can be synchronization tasks based on the use of semaphores.