Add support for adding/removing instructions to a bot programmable task.
This commit defines the class hierarchy that models the different types of instructions that a bot can execute in a programmable task. These types of instructions include the ability to navigate to a website, the ability to click on a web element, the ability to enter text in an input field, the ability to run a script, among others.