Decouple experiment and experiment session concepts in the model.
An experiment represents a usability test on a web application. It is made up of the protocols that define the activities of the participants within the test, the bots that can be included as participants and the sessions. Each experiment is defined as an instance of the Experiment class.

An experiment session represents an execution or run of an experiment where specific participants are assigned to execute the activities specified by the experiment protocols. Besides, a session includes the semaphores and variables required for synchronization and communication between participants during the session. Each experiment session is defined as an instance of the ExperimentSession class.