Add support  in the model for assigning participants to an experiment session.
Each participant is responsible for the execution of an experiment protocol. A participant can be a person or a bot. In either case, it is important to record information that describes the participant's experience throughout the session, such as the status, duration, and results of the tasks performed by the participant.