Add the view component to configure an experiment bot (WAExperimentBotComponent).
This view is used to present and configure the information of a bot designed for the  experiment. This information includes the name of the bot, its observations or notes, and the protocol that the bot must execute. When the user assigns  a protocol to the bot, the tasks that make up the protocol are loaded and displayed, adding to  the corresponding tasks (depending on the type) the links to configure the bot's actions for such tasks.