Add the view component to edit a bot instruction, that is, an instruction of a programmable task (WABotInstructionComponent).
This view allows the user to edit the human readable description of the instruction, as well as any specific property that depends on the instruction type (navigation, click, keyboard input, etc). For the edition of each specific instruction property, there is a child component (instance of WAExperimentValueComponent) that allows the user to configure if the value of the property is a constant vale or a variable's reference.