Add view to create an experiment session
This view allows the user to create a new experiment session. From this view, the user can:
- Edit the session notes to take observation about the session.
- Configure the session participants. That is select for each protocol available the type of user responsible for its execution (a person or one of the experiment bots).
- Set up the initial values of the semaphores and variables included in the session.

Also in this commit, the configuration view of an experiment was modified to include the section that presents the list of sessions of the experiment, along with a button to create new sessions.