Add the ability to show/hide participant task results in the detailed view of a session
In the list of session participants, a link has been added for each participant that allows to display in detail the results of his tasks. This information is displayed in a table that contains for each task assigned to the participant the following data: task number in the sequence, name, type, status, start time, duration and other properties. The user can hide the results of a participant's tasks at any time by clicking on the link to hide this information.