Add API endpoint to let a given participant to join his session
This method retrieves the participant with the given ID from the model. It then starts the experiment session associated with the participant (if it isn't already started). It starts the participant's activity in the session. It saves the changes. Finally, it returns the participant's data in JSON format. If the participant isn't found, it returns a response with an error message.