Change API endpoint to update the value (status) of the given semaphore
This method first retrieves the semaphore with the given ID for the specified session. Then, updates the status of the retrieved semaphore with the status passed as part of the payload. In case of the new value implies a signalling, this would let other participants (including bots) pass through the semaphore. Finally, returns semaphore data in JSON format.